backend/requirements.txt
gunicorn==19.9.0
amqp==5.0.6; python_version >= "3.6"
appdirs==1.4.4; python_version >= "3.6"
asgiref==3.4.1; python_version >= "3.6"
backports.entry-points-selectable==1.1.0; python_version >= "2.7" and python_full_version >= "3.6.1"
billiard==3.6.4.0; python_version >= "3.6"
black==20.8b1; python_version >= "3.6"
boto3==1.18.58; python_version >= "3.6"
botocore==1.21.58; python_version >= "3.6"
cached-property==1.5.2; python_version < "3.8" and python_version >= "3.6"
celery==5.1.2; python_version >= "3.6"
certifi==2021.10.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
cffi==1.14.6; python_version >= "3.7"
cfgv==3.3.1; python_full_version >= "3.6.1"
charset-normalizer==2.0.6; python_full_version >= "3.6.0" and python_version >= "3"
click-didyoumean==0.3.0; python_full_version >= "3.6.2" and python_full_version < "4.0.0" and python_version >= "3.6"
click-plugins==1.1.1; python_version >= "3.6"
click-repl==0.2.0; python_version >= "3.6"
click==7.1.2; python_full_version >= "3.6.2" and python_version >= "3.6" and python_full_version < "4.0.0" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6")
coreapi==2.3.3; python_version >= "3.6"
coreschema==0.0.4; python_version >= "3.6"
defusedxml==0.7.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_version >= "3.6" and python_full_version >= "3.5.0"
diff-match-patch==20200713; python_version >= "3.6"
distlib==0.3.3; python_full_version >= "3.6.1"
dj-rest-auth==2.1.5; python_version >= "3.5"
django-admin-autocomplete-filter==0.7.1
django-allauth==0.44.0
django-auditlog==1.0a1
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.10.0; python_version >= "3.6"
django-debug-toolbar==3.2.2; python_version >= "3.6"
django-ebhealthcheck==1.0.1
django-filter==2.4.0; python_version >= "3.5"
django-import-export==2.6.1; python_version >= "3.6"
django-phonenumber-field==6.4.0
django-two-factor-auth==1.15.2
django-otp==1.2.1
django-jsonfield==1.4.1
django-s3-storage==0.13.4
django-simple-history==3.0.0; python_version >= "3.6"
django-timezone-field==4.2.1; python_version >= "3.5"
django==3.2.8; python_version >= "3.6"
djangorestframework-simplejwt==4.8.0; python_version >= "3.7"
djangorestframework==3.12.4; python_version >= "3.5"
drf-nested-routers==0.92.5; python_version >= "3.5"
drf-yasg==1.20.0; python_version >= "3.6"
et-xmlfile==1.1.0; python_version >= "3.6"
factory-boy==3.2.0; python_version >= "3.6"
faker==9.2.0; python_version >= "3.6"
filelock==3.3.0; python_version >= "3.6" and python_full_version >= "3.6.1"
flake8==3.9.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
identify==2.3.0; python_full_version >= "3.6.1"
idna==3.2; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
importlib-metadata==4.8.1; python_version < "3.8" and python_full_version >= "3.6.1" and python_version >= "3.6" and (python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.8" or python_full_version >= "3.5.0" and python_version < "3.8" and python_version >= "3.6")
inflection==0.5.1; python_version >= "3.6"
itypes==1.2.0; python_version >= "3.6"
jinja2==3.0.2; python_version >= "3.6"
jmespath==0.10.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
kombu==5.1.0; python_version >= "3.6"
markuppy==1.14; python_version >= "3.6"
markupsafe==2.0.1; python_version >= "3.6"
mccabe==0.6.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
mypy-extensions==0.4.3; python_version >= "3.6"
nodeenv==1.6.0; python_full_version >= "3.6.1"
oauthlib==3.1.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
odfpy==1.4.1; python_version >= "3.6"
openpyxl==3.0.9; python_version >= "3.6"
packaging==21.0; python_version >= "3.6"
pathspec==0.9.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
platformdirs==2.4.0; python_version >= "3.6" and python_full_version >= "3.6.1"
pre-commit==2.15.0; python_full_version >= "3.6.1"
prompt-toolkit==3.0.20; python_full_version >= "3.6.2" and python_version >= "3.6"
psycopg2-binary==2.9.1; python_version >= "3.6"
phonenumbers==8.13.13
pycodestyle==2.7.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pycparser==2.20; python_version >= "3.7" and python_full_version < "3.0.0" or python_version >= "3.7" and python_full_version >= "3.4.0"
pyflakes==2.3.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pyjwt[crypto]==2.0.0; python_version >= "3.7"
pyparsing==2.4.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
python-crontab==2.5.1
python-dateutil==2.8.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
python3-openid==3.2.0
pytz==2021.3; python_version >= "3.6"
pyyaml==5.4.1; python_version >= "3.6" and python_full_version >= "3.6.1"
pyotp==2.8.0
redis==3.5.3; python_version >= "3.6" and python_full_version < "3.0.0" or python_version >= "3.6" and python_full_version >= "3.5.0"
regex==2021.10.8; python_version >= "3.6"
requests-oauthlib==1.3.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
requests==2.26.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
rollbar==0.15.2
reportlab==3.6.12
ruamel.yaml.clib==0.2.6; platform_python_implementation == "CPython" and python_version < "3.10" and python_version >= "3.6"
ruamel.yaml==0.17.16; python_version >= "3.6"
s3transfer==0.5.0; python_version >= "3.6"
six==1.16.0; python_full_version >= "3.6.1" and python_version >= "3.6"
sqlparse==0.4.2; python_version >= "3.6"
twilio==8.2.1
tablib==3.0.0; python_version >= "3.6"
text-unidecode==1.3; python_version >= "3.6"
toml==0.10.2; python_full_version >= "3.6.1" and python_version >= "3.6"
typed-ast==1.4.3; python_version >= "3.6"
typing-extensions==3.10.0.2; python_version < "3.8" and python_version >= "3.6" and python_full_version >= "3.6.1"
uritemplate==3.0.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
urllib3==1.26.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.6"
vine==5.0.0; python_version >= "3.6"
virtualenv==20.8.1; python_full_version >= "3.6.1"
wcwidth==0.2.5; python_full_version >= "3.6.2" and python_version >= "3.6"
xlrd==2.0.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_version >= "3.6" and python_full_version >= "3.6.0"
xlwt==1.3.0; python_version >= "3.6"
zipcodes==1.2.0
zipp==3.6.0; python_version < "3.8" and python_full_version >= "3.6.1" and python_version >= "3.6"
django-allow-cidr==0.5.0
python-dotenv==0.21.1